added CustomValidator.php service. Cleaned code in TagController.php. Must improve performance on context state.

+++ src/Controller/TagController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Tag;
 use App\Repository\TagRepository;
+use App\Service\CustomValidator;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -29,7 +30,7 @@ class TagController extends AbstractController
      */
     private $validator;
 
-    public function __construct(TagRepository $tagRepository, EntityManagerInterface $entityManager, ValidatorInterface $validator)
+    public function __construct(TagRepository $tagRepository, EntityManagerInterface $entityManager, CustomValidator $validator)
     {
 
         $this->tagRepository = $tagRepository;
@@ -63,29 +64,18 @@ class TagController extends AbstractController
         $tag = new Tag();
         $tag->setName($content->name);
 
-        $tagNameErrors = $this->validator->validate($tag);
+        $errors = $this->validator->validate($tag);
 
-        $errors = [];
-        if (sizeof($tagNameErrors) > 0) {
-
-//            enable this to generate an array of all errors instead of just one
-//            foreach ($tagNameErrors as $tagNameError) {
-//                $errors[] = $tagNameError->getMessage();
-//            }
-
-            $errors[] = $tagNameErrors[0]->getMessage();
-
+        if (sizeof($errors) > 0) {
             return $this->json(
                 ['message' => ['text' => $errors, 'level' => 'error']]
             );
+        } else {
+            $this->entityManager->persist($tag);
+            $this->entityManager->flush();
+
+            return $this->json(['tag' => $tag->normalize(),]);
         }
-
-        $this->entityManager->persist($tag);
-        $this->entityManager->flush();
-
-        return $this->json([
-            'tag' => $tag->normalize(),
-        ]);
     }
 
     /**
@@ -94,34 +84,23 @@ class TagController extends AbstractController
      * @param Tag $tag
      * @return JsonResponse
      */
-    public function update(Request $request, Tag $tag)
+    public
+    function update(Request $request, Tag $tag)
     {
         $content = json_decode($request->getContent());
         $tag->setName($content->name);
 
-        $this->validator->validate($tag);
+        $errors = $this->validator->validate($tag);
 
-        $validatorErrors = $this->validator->validate($tag);
-
-        $errors = [];
-        if (sizeof($validatorErrors) > 0) {
-
-//            enable this to generate an array of all errors instead of just one
-//            foreach ($validatorErrors as $error) {
-//                $errors[] = $error->getMessage();
-//            }
-
-            $errors[] = $validatorErrors[0]->getMessage();
-
+        if (sizeof($errors) > 0) {
             return $this->json(
                 ['message' => ['text' => $errors, 'level' => 'error']]
             );
         } else {
             $this->entityManager->flush();
-
-            return $this->json([
-                ['message' => ['text' => 'Tag has been updated!', 'level' => 'success']],
-            ]);
+            return $this->json(
+                ['message' => ['text' => 'Tag has been updated!', 'level' => 'success']]
+            );
         }
     }
 
@@ -130,13 +109,21 @@ class TagController extends AbstractController
      * @param Tag $tag
      * @return JsonResponse
      */
-    public function delete(Tag $tag)
+    public
+    function delete(Tag $tag)
     {
+        $errors = $this->validator->validate($tag);
+        if (sizeof($errors) > 0) {
+            return $this->json([
+                ['message' => ['text' => $errors, 'level' => 'error']]
+            ]);
+        }
+
         $this->entityManager->remove($tag);
         $this->entityManager->flush();
 
-        $form = $this->createFormBuilder();
-
-        return $this->json([]);
+        return $this->json(
+            ['message' => ['text' => 'Tag has been deleted!', 'level' => 'success']]
+        );
     }
 }
